refactor: 모임 생성 퍼널 코드 스플리팅
